Update unit tests

+++ src/test/java/sh/casey/subtitler/shifter/AssSubtitleShifterTest.java
@@ -24,7 +24,7 @@ public class AssSubtitleShifterTest {
     public void testShift() {
         final Long after = 728000L;
         final Long before = 780000L;
-        final Integer shift = 12480;
+        final Long shift = 12480L;
         final String filePath = "src/test/resources/ass/[Judas] Bleach - S01E04 - 004.en.ass";
         final AssSubtitleFile file = reader.read(filePath);
         final AssSubtitleFile shifted = reader.read(filePath);
